Made auth more flexible - not generating pages & routes anymore, just the forms.
