+++ examples/realworld/ext/user/components/UserSettingsPage.js
@@ -13,7 +13,7 @@ const UserSettingsPage = () => {
   // TODO: Instead of this logic here, I wish I could use ACL via Wasp and just
   //   receive user via props instead of useAuth().
   if (!user || isError) {
-    return <span> Please <Link to='/signin'>log in</Link>. </span>
+    return <span> Please <Link to='/login'>log in</Link>. </span>
   }
 
   return (
