+++ examples/realworld/main.wasp
@@ -14,11 +14,7 @@ page Main {
   component: import Main from "@ext/MainPage.js"
 }
 
-// TODO: I had to change route to '/signin' because Wasp right now occupies
-//   '/login' with its own login page. By RWA spec, url should be '/login'.
-//   We should change this back to '/login' once we make Wasp more flexible
-//   in this regard.
-route "/signin" -> page LogIn
+route "/login" -> page LogIn
 page LogIn {
   component: import LogIn from "@ext/auth/LoginPage.js"
 }
