+++ examples/tutorials/TodoApp/ext/LoginPage.js
@@ -1,0 +1,18 @@
+import React, { useState } from 'react'
+import { useHistory, Link } from 'react-router-dom'
+
+import LoginForm from '@wasp/auth/forms/Login'
+
+const LoginPage = (props) => {
+  return (
+    <>
+      <LoginForm/>
+      <br/>
+      <span>
+        I don't have an account yet (<Link to="/signup">go to signup</Link>).
+      </span>
+    </>
+  )
+}
+
+export default LoginPage
