+++ examples/tutorials/TodoApp/ext/SignupPage.js
@@ -1,0 +1,18 @@
+import React, { useState } from 'react'
+import { useHistory, Link } from 'react-router-dom'
+
+import SignupForm from '@wasp/auth/forms/Signup'
+
+const SignupPage = (props) => {
+  return (
+    <>
+      <SignupForm/>
+      <br/>
+      <span>
+        I already have an account (<Link to="/login">go to login</Link>).
+      </span>
+    </>
+  )
+}
+
+export default SignupPage
