+++ waspc/src/Generator/WebAppGenerator/RouterGenerator.hs
@@ -2,7 +2,6 @@ module Generator.WebAppGenerator.RouterGenerator
     ( generateRouter
     ) where
 
-import           Data.Maybe                       (isJust)
 import           Data.Aeson                       (ToJSON (..), object, (.=))
 import qualified Path                             as P
 
@@ -56,49 +55,9 @@ createRouterTemplateData wasp = RouterTemplateData
     , _pagesToImport = pages
     }
     where
-        maybeAuth = Wasp.getAuth wasp
-        routes = (Wasp.getRoutes wasp) ++ (if isJust maybeAuth then authRoutes else [])
+        routes = Wasp.getRoutes wasp
+        pages = map createPageTemplateData $ Wasp.getPages wasp
 
-        -- TODO(matija): It would be nicer if we were changing AST "higher" in the program, e.g.
-        -- adding built-in pages rather than doing it here in the generator -> that way we'd keep
-        -- generator code simpler and push the logic higher.
-        pages = (map createPageTemplateData $ Wasp.getPages wasp) ++
-                 (if isJust maybeAuth then authPages else [])
-
-        authRoutes :: [Wasp.Route.Route]
-        authRoutes =
-            [ Wasp.Route.Route -- Signup route
-                { Wasp.Route._urlPath = signupPageRoute
-                , Wasp.Route._targetPage = signupPageName
-                }
-            , Wasp.Route.Route -- Login route
-                { Wasp.Route._urlPath = loginPageRoute
-                , Wasp.Route._targetPage = loginPageName
-                }
-            ]
-
-        authPages :: [PageTemplateData]
-        authPages =
-            [ PageTemplateData -- Signup page
-                { _importWhat = signupPageName
-                , _importFrom =
-                    "./" ++ (SP.fromRelFileP $
-                             SP.fromPathRelFileP [P.relfile|auth/pages/Signup.js|])
-                }
-            , PageTemplateData -- Login page
-                { _importWhat = loginPageName
-                , _importFrom =
-                    "./" ++ (SP.fromRelFileP $
-                             SP.fromPathRelFileP [P.relfile|auth/pages/Login.js|])
-                }
-            ]
-
-        signupPageName = "Signup"
-        signupPageRoute = "/signup"
-
-        loginPageName = "Login"
-        loginPageRoute = "/login"
-        
 
 createPageTemplateData :: Wasp.Page.Page -> PageTemplateData
 createPageTemplateData page = PageTemplateData
